BSAPP-554 Ausbesserung bzw Vereinigung der beiden Wettkampfergebnis-Services
