+++ bogenliga/src/assets/i18n/de.json
@@ -47,7 +47,7 @@
   "REGIONEN": {
     "REGIONEN": {
       "TITLE": "Regionen",
-      "DESCRIPTION": "Um eine Liga auszuwählen klicken Sie bitte auf die entsprechende Region. Z.b. für die Württembergliga auf Württembergischer Landesverband"
+      "DESCRIPTION": "Um eine Liga auszuwählen klicken Sie bitte auf die entsprechende Region, beispielsweise für die Württembergliga auf den Württembergischer Schützenverband."
     }
   },
   "VEREINE": {
